+++ app/src/main/java/com/anytypeio/anytype/di/feature/NavigationDI.kt
@@ -2,6 +2,7 @@ package com.anytypeio.anytype.di.feature
 
 import com.anytypeio.anytype.analytics.base.Analytics
 import com.anytypeio.anytype.core_utils.di.scope.PerScreen
+import com.anytypeio.anytype.domain.`object`.ObjectTypesProvider
 import com.anytypeio.anytype.domain.block.repo.BlockRepository
 import com.anytypeio.anytype.domain.config.GetConfig
 import com.anytypeio.anytype.domain.misc.UrlBuilder
@@ -43,13 +44,15 @@ object PageNavigationModule {
         urlBuilder: UrlBuilder,
         getObjectInfoWithLinks: GetObjectInfoWithLinks,
         getConfig: GetConfig,
-        analytics: Analytics
+        analytics: Analytics,
+        objectTypesProvider: ObjectTypesProvider
     ): PageNavigationViewModelFactory =
         PageNavigationViewModelFactory(
             urlBuilder = urlBuilder,
             getObjectInfoWithLinks = getObjectInfoWithLinks,
             getConfig = getConfig,
-            analytics = analytics
+            analytics = analytics,
+            objectTypesProvider = objectTypesProvider
         )
 
     @JvmStatic