+++ presentation/src/main/java/com/anytypeio/anytype/presentation/moving/MoveToViewModelFactory.kt
@@ -2,6 +2,7 @@ package com.anytypeio.anytype.presentation.moving
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.anytypeio.anytype.domain.`object`.ObjectTypesProvider
 import com.anytypeio.anytype.domain.block.interactor.Move
 import com.anytypeio.anytype.domain.config.GetConfig
 import com.anytypeio.anytype.domain.misc.UrlBuilder
@@ -11,7 +12,8 @@ class MoveToViewModelFactory(
     private val urlBuilder: UrlBuilder,
     private val getObjectInfoWithLinks: GetObjectInfoWithLinks,
     private val getConfig: GetConfig,
-    private val move: Move
+    private val move: Move,
+    private val objectTypesProvider: ObjectTypesProvider
 ) : ViewModelProvider.Factory {
 
     @Suppress("UNCHECKED_CAST")
@@ -20,7 +22,8 @@ class MoveToViewModelFactory(
             urlBuilder = urlBuilder,
             getObjectInfoWithLinks = getObjectInfoWithLinks,
             getConfig = getConfig,
-            move = move
+            move = move,
+            objectTypesProvider = objectTypesProvider
         ) as T
     }
 }