+++ presentation/src/main/java/com/anytypeio/anytype/presentation/navigation/PageNavigationViewModel.kt
@@ -9,6 +9,7 @@ import com.anytypeio.anytype.core_utils.common.EventWrapper
 import com.anytypeio.anytype.core_utils.ext.timber
 import com.anytypeio.anytype.core_utils.ui.ViewState
 import com.anytypeio.anytype.core_utils.ui.ViewStateViewModel
+import com.anytypeio.anytype.domain.`object`.ObjectTypesProvider
 import com.anytypeio.anytype.domain.config.GetConfig
 import com.anytypeio.anytype.domain.misc.UrlBuilder
 import com.anytypeio.anytype.domain.page.navigation.GetObjectInfoWithLinks
@@ -22,9 +23,9 @@ class PageNavigationViewModel(
     private val urlBuilder: UrlBuilder,
     private val getObjectInfoWithLinks: GetObjectInfoWithLinks,
     private val getConfig: GetConfig,
-    private val analytics: Analytics
-) :
-    ViewStateViewModel<ViewState<PageNavigationView>>(),
+    private val analytics: Analytics,
+    private val objectTypesProvider: ObjectTypesProvider
+) : ViewStateViewModel<ViewState<PageNavigationView>>(),
     SupportNavigation<EventWrapper<AppNavigation.Command>> {
 
     private var pageId: String = ""
@@ -56,8 +57,8 @@ class PageNavigationViewModel(
                                     subtitle = documentInfo.snippet.orEmpty(),
                                     image = documentInfo.obj.getImagePath(urlBuilder),
                                     emoji = documentInfo.obj.getEmojiPath(),
-                                    inbound = links.inbound.map { it.toView(urlBuilder) },
-                                    outbound = links.outbound.map { it.toView(urlBuilder) }
+                                    inbound = links.inbound.map { it.toView(urlBuilder, objectTypesProvider.get()) },
+                                    outbound = links.outbound.map { it.toView(urlBuilder, objectTypesProvider.get()) }
                                 )
                             )
                         )