+++ presentation/src/main/java/com/anytypeio/anytype/presentation/navigation/PageNavigationViewModelFactory.kt
@@ -3,6 +3,7 @@ package com.anytypeio.anytype.presentation.navigation
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.anytypeio.anytype.analytics.base.Analytics
+import com.anytypeio.anytype.domain.`object`.ObjectTypesProvider
 import com.anytypeio.anytype.domain.config.GetConfig
 import com.anytypeio.anytype.domain.misc.UrlBuilder
 import com.anytypeio.anytype.domain.page.navigation.GetObjectInfoWithLinks
@@ -11,7 +12,8 @@ class PageNavigationViewModelFactory(
     private val urlBuilder: UrlBuilder,
     private val getObjectInfoWithLinks: GetObjectInfoWithLinks,
     private val getConfig: GetConfig,
-    private val analytics: Analytics
+    private val analytics: Analytics,
+    private val objectTypesProvider: ObjectTypesProvider
 ) : ViewModelProvider.Factory {
 
     @Suppress("UNCHECKED_CAST")
@@ -20,7 +22,8 @@ class PageNavigationViewModelFactory(
             urlBuilder = urlBuilder,
             getObjectInfoWithLinks = getObjectInfoWithLinks,
             getConfig = getConfig,
-            analytics = analytics
+            analytics = analytics,
+            objectTypesProvider = objectTypesProvider
         ) as T
     }
 }